Ability to load initial TPCC data to YDB
* YDB DDL
* Use Double instead of Decimal and Float

+++ src/main/java/com/oltpbenchmark/benchmarks/tpcc/TPCCUtil.java
@@ -50,10 +50,10 @@ public class TPCCUtil {
         c.c_zip = rs.getString("c_zip");
         c.c_phone = rs.getString("c_phone");
         c.c_credit = rs.getString("c_credit");
-        c.c_credit_lim = rs.getFloat("c_credit_lim");
-        c.c_discount = rs.getFloat("c_discount");
-        c.c_balance = rs.getFloat("c_balance");
-        c.c_ytd_payment = rs.getFloat("c_ytd_payment");
+        c.c_credit_lim = rs.getDouble("c_credit_lim");
+        c.c_discount = rs.getDouble("c_discount");
+        c.c_balance = rs.getDouble("c_balance");
+        c.c_ytd_payment = rs.getDouble("c_ytd_payment");
         c.c_payment_cnt = rs.getInt("c_payment_cnt");
         c.c_since = rs.getTimestamp("c_since");
         return c;
