+++ src/main/java/com/oltpbenchmark/benchmarks/tpcc/pojo/History.java
@@ -28,7 +28,7 @@ public class History {
     public int h_d_id;
     public int h_w_id;
     public Timestamp h_date;
-    public float h_amount;
+    public double h_amount;
     public String h_data;
 
     @Override
