+++ src/main/java/com/oltpbenchmark/benchmarks/tpcc/pojo/OrderLine.java
@@ -30,7 +30,7 @@ public class OrderLine {
     public int ol_supply_w_id;
     public int ol_quantity;
     public Timestamp ol_delivery_d;
-    public float ol_amount;
+    public double ol_amount;
     public String ol_dist_info;
 
     @Override
